refactor(step3): 불필요한 서비스 레이어 제거
