--- src/main/kotlin/lotto/component/LottoInputConverter.kt
@@ -1,41 +1,0 @@
-package lotto.component
-
-import lotto.model.LottoNumbers
-import lotto.model.WinningNumbers
-
-class LottoInputConverter(
-    private val lottoInputValidator: LottoInputValidator
-) {
-    fun getPurchasePrice(purchasePrice: String?): Int {
-        return purchasePrice
-            .run { lottoInputValidator.validatePurchasePrice(this) }
-            .toInt()
-    }
-
-    fun getLottoNumbers(purchasePrice: Int): List<LottoNumbers> {
-        val lottoNumbersCount: Int = lottoInputValidator.validateLottoNumbersCount(purchasePrice / Lotto.LOTTO_PRICE)
-
-        return LottoNumbers.generate(lottoNumbersCount)
-    }
-
-    fun getBonusNumber(bonusNumber: String?): Int {
-        return lottoInputValidator
-            .validateLottoNumber(bonusNumber)
-            .toInt()
-    }
-
-    fun getWinningNumbers(winningNumbers: String?): WinningNumbers {
-        val numbers = winningNumbers
-            .run { lottoInputValidator.validateLottoNumbers(this) }
-            .split(LOTTO_NUMBERS_SEPARATOR)
-            .map { lottoInputValidator.validateLottoNumber(it) }
-            .map { it.toInt() }
-            .run { lottoInputValidator.validateLottoNumberCount(this) }
-
-        return WinningNumbers.create(numbers)
-    }
-
-    companion object {
-        const val LOTTO_NUMBERS_SEPARATOR = ", "
-    }
-}
