+++ src/main/kotlin/lotto/controller/LottoViewController.kt
@@ -1,6 +1,7 @@
 package lotto.controller
 
-import lotto.component.LottoInputConverter
+import lotto.component.Lotto
+import lotto.component.LottoInputValidator
 import lotto.model.LottoNumbers
 import lotto.model.LottoResult
 import lotto.model.WinningNumbers
@@ -10,28 +11,58 @@ import lotto.view.LottoResultView
 
 class LottoViewController(
     private val lottoInputView: LottoInputView,
-    private val lottoInputConverter: LottoInputConverter,
+    private val lottoInputValidator: LottoInputValidator,
     private val lottoResultView: LottoResultView
 ) {
     fun getPurchasePrice(): Int {
         val purchasePrice: String? = lottoInputView.getPurchasePrice()
 
-        return lottoInputConverter.getPurchasePrice(purchasePrice)
+        return convertPurchasePrice(purchasePrice)
     }
 
     fun getLottoNumbers(purchasePrice: Int): List<LottoNumbers> {
-        return lottoInputConverter.getLottoNumbers(purchasePrice)
+        return convertLottoNumbers(purchasePrice)
     }
 
     fun getBonusNumber(): Int {
         val bonusNumber: String? = lottoInputView.getBonusNumber()
 
-        return lottoInputConverter.getBonusNumber(bonusNumber)
+        return convertBonusNumber(bonusNumber)
     }
 
     fun getWinningNumbers(): WinningNumbers {
         val winningNumbers: String? = lottoInputView.getWinningNumbers()
-        return lottoInputConverter.getWinningNumbers(winningNumbers)
+
+        return convertWinningNumbers(winningNumbers)
+    }
+
+    private fun convertPurchasePrice(purchasePrice: String?): Int {
+        return purchasePrice
+            .run { lottoInputValidator.validatePurchasePrice(this) }
+            .toInt()
+    }
+
+    private fun convertLottoNumbers(purchasePrice: Int): List<LottoNumbers> {
+        val lottoNumbersCount: Int = lottoInputValidator.validateLottoNumbersCount(purchasePrice / Lotto.LOTTO_PRICE)
+
+        return LottoNumbers.generate(lottoNumbersCount)
+    }
+
+    private fun convertBonusNumber(bonusNumber: String?): Int {
+        return lottoInputValidator
+            .validateLottoNumber(bonusNumber)
+            .toInt()
+    }
+
+    private fun convertWinningNumbers(winningNumbers: String?): WinningNumbers {
+        val numbers = winningNumbers
+            .run { lottoInputValidator.validateLottoNumbers(this) }
+            .split(LOTTO_NUMBERS_SEPARATOR)
+            .map { lottoInputValidator.validateLottoNumber(it) }
+            .map { it.toInt() }
+            .run { lottoInputValidator.validateLottoNumberCount(this) }
+
+        return WinningNumbers.create(numbers)
     }
 
     fun printPurchasedLottoNumbers(lottoNumbers: List<LottoNumbers>) {
@@ -40,5 +71,9 @@ class LottoViewController(
 
     fun printLottoResult(lottoResult: LottoResult) {
         lottoResultView.printLottoResult(lottoResult)
+    }
+
+    companion object {
+        const val LOTTO_NUMBERS_SEPARATOR = ", "
     }
 }
